# Conflicts:
#	app/src/main/java/com/starters/yeogida/data/api/PlaceService.kt

+++ app/src/main/java/com/starters/yeogida/data/local/PlaceMapData.kt
@@ -3,7 +3,7 @@ package com.starters.yeogida.data.local
 data class PlaceMapData(
     val placeThumbnail: String,
     val placeName: String,
-    val star: Double,
+    val star: Float,
     val comment: Int,
     val tag: String,
     val address: String
